Fix error when creating ActionController::TestRequest
When building a full view context, Draper called
ActionController::TestRequest.new with no parameters. Since rails 5,
that initializer needs to be called with 2 parameters, an env and a
session. However, there is also a new 'create' constructor that takes no
args and creates a TestRequest with the default env values.
When using this new 'create' constructor, rake tasks in an app including
draper will run again.

This is still a temporary fix; in rails 5.1 ActionController::TestCase
will be deprecated and moved to a gem. At this point either this will
have to be removed or the new gem will become a dependency.